Hide implementation better
